requirements.txt
asgiref==3.5.2
Django==4.1.1
djangorestframework==3.14.0
flake8==5.0.4
mccabe==0.7.0
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.21.0
pytz==2022.2.1
sqlparse==0.4.3
